dom/base/test/test_document_importNode_document.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=1177914
-->
<head>
  <title>Test for Bug 1177914</title>
  <script type="text/javascript" src="/MochiKit/MochiKit.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1177914">Mozilla Bug 1177914</a>
<p id="display"></p>
<div id="content" style="display: none">
</div>
<pre id="test">
<script class="testbody" type="text/javascript">
var thrownException = false;

try {
  document.importNode(document);
} catch(err) {
  thrownException = err;
}

ok(thrownException !== false, "An exception should've been thrown");
is(thrownException.name, "NotSupportedError", "A NotSupportedError exception should've been thrown");
</script>
</pre>
</body>
</html>
